+++ core/src/main/java/dev/nandi0813/practice/manager/gui/guis/queue/RankedGui.java
@@ -14,7 +14,7 @@ public class RankedGui extends QueueSelectorGui {
     }
 
     @Override
-    protected int getUpdateCooldownMinutes() {
+    protected long getUpdateCooldownMinutes() {
         return ConfigManager.getInt("QUEUE.RANKED.GUI-UPDATE-MINUTE");
     }
